Created a new track collision algorithm, added soem changed to the lightin, added the the logic for the car

+++ src/tema2/3dGeometries/GeometryObject.cpp
@@ -71,7 +71,10 @@ void GeometryObject::Render(
 
 	// Send the positions
 	location = glGetUniformLocation(m_shader->program, "light_positions");
-	glUniform3fv(location, lightingPositions.size(), reinterpret_cast<GLfloat*>(lightingPositions.data()));
+	glUniform3fv(
+		location,
+		static_cast<GLint>(lightingPositions.size()),
+		reinterpret_cast<GLfloat*>(lightingPositions.data()));
 
 	m_mesh->Render();
 }
